Prevent metrics panic from holding lock
If IncLabel or Trace are called without first calling EntityType, they will panic because the groupEntityMetrics.em is nil. Previously this panic would occur while holding m.Lock, causing the lock to never be released and blocking all future metrics calls.

Now we always defer m.Unlock(), so that even if there is a panic, we are certain to release the lock.