Change wrapper handling for CDC
CDC requests previously went through requestWrapper(...). This function adds middleware, but was complicated because CDC does not use the same middleware as a regular read/write request. In particular, CDC should not have request timeouts (since CDC connections are long lived) and do not have EntityType. For this reason it was cleaner to split the CDC middleware into its own cdcWrapper handler.

Auth for CDC also had to be updated. In older etre releases, CDC bypassed the authorization flow. To run CDC through authorization, we added a new operation OP_CDC to the auth package along with a boolean CDC flag in the ACL. This will require existing ACLs to be updated to allow CDC, but was cleaner than other options such as allowing CDC if the user had READ access to all known types. It also allows for users that can execute READ requests, but cannot stream CDC changes.