Fix CDC and metrics handling (#77)
CDC requests caused a panic which then never released a mutex lock within the metrics collection library. This in turn blocked all other requests. This PR includes that fix, plus a handful of improvements to harden against this sort of issue in the future.

* Change wrapper handling for CDC

CDC requests previously went through requestWrapper(...). This function adds middleware, but was complicated because CDC does not use the same middleware as a regular read/write request. In particular, CDC should not have request timeouts (since CDC connections are long lived) and do not have EntityType. For this reason it was cleaner to split the CDC middleware into its own cdcWrapper handler.

Auth for CDC also had to be updated. In older etre releases, CDC bypassed the authorization flow. To run CDC through authorization, we added a new operation OP_CDC to the auth package along with a boolean CDC flag in the ACL. This will require existing ACLs to be updated to allow CDC, but was cleaner than other options such as allowing CDC if the user had READ access to all known types. It also allows for users that can execute READ requests, but cannot stream CDC changes.

* Prevent metrics panic from holding lock with `defer m.Unlock()`

If IncLabel or Trace are called without first calling EntityType, they will panic because the groupEntityMetrics.em is nil. Previously this panic would occur while holding m.Lock, causing the lock to never be released and blocking all future metrics calls.

Now we always defer m.Unlock(), so that even if there is a panic, we are certain to release the lock.

* API tests now use real metrics

This bug was not caught in the existing tests which used a mock metrics collector. The mock collector did not panic the way the real collector did. Now the tests in the api package use the real metrics collector, giving higher fidelity tests in the future.