Add _created and _updated metalables (#89)
The code change itself is relatively small and straightforward:
* Add the new metalable in etre.go
* Add the validations in validate.go
* Maintain the label in store.go
* Update the CLI to pretty-print in a human readable form (instead of the raw Unix nano time)

The test changes were more involved, particularly because the timestamps are not deterministic. Most of the tests that validate actual data now follow a pattern where they put the actual values into the expected entity for the assert.Equal(t, expected, actual), and then do a separate check that the actual time is a realistic value (in the last few seconds).

Also fixed a bug with Entity.Rev(). It expected an int type, but json.Unmarshal sets it to a float64 which caused a panic. Now that is fixed to allow the various int flavors _and_ float.